auth api update with device id

+++ app/Http/Controllers/API/Auth/AuthController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\API\Auth;
 
 use App\Http\Controllers\Controller;
+use App\Models\BlockUser;
 use App\Models\User;
 use App\Models\VerificationCode;
 use Carbon\Carbon;
@@ -16,19 +17,29 @@ class AuthController extends Controller
 {
     public function register(Request $request)
     {
-        $data = $request->all();
-        $validator = Validator::make($data, [
+        // $data = $request->all();
+        // $validator = Validator::make($data, [
+        //     'name' => ['required', 'string', 'max:255'],
+        //     'mobile' => ['required', 'string', 'max:15', 'unique:users'],
+        //     'password' => ['required', 'string', 'min:4'],
+        //     'device_id' => ['required'],
+        // ],[
+        //     'device_id.required' => 'Device address not foound. Update app or Contact with helpline.',
+        // ]);
+        $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'mobile' => ['required', 'string', 'max:15', 'unique:users'],
             'password' => ['required', 'string', 'min:4'],
+            'device_id' => ['required'],
+        ], [
+            'device_id.required' => 'Device address not foound. Update app or Contact with helpline.',
         ]);
-        if ($validator->fails()) {
-            return response()->json($validator->errors(),422);
-        }
+
         $user = User::create([
-            'name' => $data['name'],
-            'mobile' => $data['mobile'],
-            'password' => Hash::make($data['password']),
+            'name' =>$request->name,
+            'mobile' => $request->mobile,
+            'password' => Hash::make($request->password),
+            'device_id' => $request->device_id,
             'status' => false,
         ]);
 
@@ -57,17 +68,35 @@ class AuthController extends Controller
 
         $request->validate([
             'mobile'         => 'required',
-            'password'      => 'required'
+            'password'      => 'required',
+            'device_id'      => 'required'
+        ], [
+            'device_id.required' => 'Device address not foound. Update app or Contact with helpline.',
         ]);
 
         //
         try {
             // if (Auth::attempt($request->only('mobile', 'password'))) {
+            $checkblock = BlockUser::firstWhere('device_id', $request->device_id);
+            // return $checkblock;
+            if ($checkblock) {
+                return response()->json([
+                    'message' => 'Your mobile has been block.',
+                ]);
+            }
             if (Auth::attempt(['mobile' => $request->mobile, 'password' => $request->password, 'status' => 1])) {
                 $user = Auth::user();
                 $user->tokens()->delete();
                 $token = $user->createToken(uniqid())->plainTextToken;
                 $user['token'] = $token;
+
+                $checkdevice = $user->device_id;
+                // empty($checkdevice)?? User::firstWhere('id', $user->id)->update(['device_id'=> $request->device_id]);
+                if ($checkdevice == null) {
+                    User::firstWhere('id', $user->id)->update(['device_id' => $request->device_id]);
+                }
+
+
                 return response()->json([
                     'message' => 'successfully login',
                     'token' => $token,
